Надо же - заработало хитрость в замене url на var urlNew в GlideImageLoader
