+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/userrepo/IGithubRepositoriesRepo.kt
@@ -1,4 +1,4 @@
-package com.example.popularlibs_homrworks.model.repositories.reporepos
+package com.example.popularlibs_homrworks.model.repositories.userrepo
 
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.entity.GithubUserRepos
