+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/userrepo/cashrepos/IRoomRepositiriesRepoCash.kt
@@ -1,4 +1,4 @@
-package com.example.popularlibs_homrworks.model.repositories.reporepos.cashrepos
+package com.example.popularlibs_homrworks.model.repositories.userrepo.cashrepos
 
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.entity.GithubUserRepos
