+++ app/src/main/java/com/example/popularlibs_homrworks/view/adapters/users/UsersRVAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.popularlibs_homrworks.R
-import com.example.popularlibs_homrworks.model.repositories.glide.IImageLoader
+import com.example.popularlibs_homrworks.model.glide.IImageLoader
 import com.example.popularlibs_homrworks.presenters.users.IUserListPresenter
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.item_user.view.*
